Trait pallet_linkable_tree_rpc_runtime_api::LinkableTreeApi

pub trait LinkableTreeApi<Block: BlockT, C, E, L>: Core<Block>where
    C: Encode + Decode,
    E: ElementTrait,
    L: Encode + Decode,{
    fn get_neighbor_roots(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        tree_id: u32
    ) -> Result<Vec<E>, ApiError> { ... }
    fn get_neighbor_roots_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        tree_id: u32
    ) -> Result<Vec<E>, ApiError> { ... }
    fn get_neighbor_edges(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        tree_id: u32
    ) -> Result<Vec<EdgeMetadata<C, E, L>>, ApiError> { ... }
    fn get_neighbor_edges_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        tree_id: u32
    ) -> Result<Vec<EdgeMetadata<C, E, L>>, ApiError> { ... }
}

Provided Methods

fn get_neighbor_roots(
&self,
__runtime_api_at_param__: &BlockId<Block>,
tree_id: u32
) -> Result<Vec<E>, ApiError>

Get the neighbor roots including the roots for default (empty) edges

fn get_neighbor_roots_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
tree_id: u32
) -> Result<Vec<E>, ApiError>

Get the neighbor roots including the roots for default (empty) edges

fn get_neighbor_edges(
&self,
__runtime_api_at_param__: &BlockId<Block>,
tree_id: u32
) -> Result<Vec<EdgeMetadata<C, E, L>>, ApiError>

Get the neighbor edge metadata including the metadata for default (empty) edges

fn get_neighbor_edges_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
tree_id: u32
) -> Result<Vec<EdgeMetadata<C, E, L>>, ApiError>

Get the neighbor edge metadata including the metadata for default (empty) edges